Final upload

+++ Exercise 2/run.sh
@@ -1,4 +1,7 @@
-rm feedback
+#If file "feedback" exists delete it.
+if [ -f feedback ]; then
+    rm feedback
+fi
 
 make sensor -s && \
 mkfifo feedback && \
